Blink {GAPIT3}	R Documentation

Blink

Description

Blink

Usage

Blink(
  Y = NULL,
  QTN.position = NULL,
  GD = NULL,
  GM = NULL,
  CV = NULL,
  DPP = 1e+08,
  kinship.algorithm = "FARM-CPU",
  file.output = TRUE,
  cutOff = 0.01,
  method.GLM = "FarmCPU.LM",
  Prior = NULL,
  ncpus = 1,
  maxLoop = 10,
  LD = 0.7,
  threshold.output = 1e-04,
  alpha = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1),
  WS = c(1, 1000, 10000, 1e+05, 1e+06, 1e+07),
  GP = NULL,
  FDRcut = FALSE,
  maxOut = 10,
  converge = 1,
  iteration.output = FALSE,
  acceleration = 0,
  threshold = NA,
  model = "A",
  MAF.calculate = FALSE,
  plot.style = "FarmCPU",
  p.threshold = NA,
  maf.threshold = 0,
  bound = FALSE,
  method.sub = "reward",
  method.sub.final = "reward",
  stepwise = FALSE,
  BIC.method = "naive",
  LD.wise = FALSE,
  time.cal = FALSE,
  Prediction = FALSE
)

Arguments

Y
= NULL, data.frame of phenotypic data, column one is sample (taxa), column two is a trait, subsequent columns are other traits.

QTN.position
= NULL,

GD
= NULL, data.frame of genetic data in 'numerical' format, samples in rows, variants in columns.

GM
= NULL, Genetic Map data.frame to provide genomic coordinates for GD

CV
= NULL, Covariates

DPP
= 100000000,

kinship.algorithm
= "FARM-CPU",

file.output
= TRUE,

cutOff
= 0.01,

method.GLM
= "FarmCPU.LM",

Prior
= NULL,

ncpus
= 1,

maxLoop
= 10,

LD
= 0.7,

threshold.output
= .0001,

alpha
= c(.01,.05,.1,.2,.3,.4,.5,.6,.7,.8,.9,1),

WS
= c(1e0,1e3,1e4,1e5,1e6,1e7),

GP
= NULL,

FDRcut
= FALSE,

maxOut
= 10,

converge
= 1,

iteration.output
= FALSE,

acceleration
= 0,

threshold
= NA,

model
= "A",

MAF.calculate
= FALSE,

plot.style
= "FarmCPU",

p.threshold
= NA,

maf.threshold
= 0,

bound
= FALSE,

method.sub
= "reward",

method.sub.final
= "reward",

stepwise
= FALSE,

BIC.method
= "naive",

LD.wise
= FALSE,

time.cal
= FALSE,

Prediction
= F

Value

list(GWAS=GWAS,myGLM=myGLM,PEV = PEV,seqQTN=seqQTN)

[Package GAPIT3 version 3.1.0 Index]